Add prop to book
