+++ package.json
@@ -9,10 +9,12 @@
     "babel-cli": "^6.26.0",
     "babel-preset-react-app": "^3.1.2",
     "big.js": "^6.2.1",
-    "react-router-dom": "^6.8.1",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
+    "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
